more reorg progress
