+++ src/libraries/TRACKING/DTrackFinder.h
@@ -127,7 +127,7 @@ class DTrackFinder:public jana::JObject{
             const DVector3 &origin,DVector3 *poca=NULL) const;
       double FindDoca(const DVector3 &pos1,const DVector3 &mom1,
             const DVector3 &pos2,const DVector3 &mom2,
-            DVector3 *poca) const;
+            DVector3 *poca=NULL) const;
       bool FindIntersectionsWithCylinder(double R,const DVector3 &dir,
             const DVector3 &pos,DVector3 &out1,
             DVector3 &out2) const;
